requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
google-api-python-client>=2.116.0
openai>=1.12.0
python-telegram-bot>=20.7
aiosqlite>=0.19.0
httpx>=0.25.2
apscheduler>=3.10.4
streamlit>=1.41.0
